requirements.txt
Flask==2.3.3
tensorflow>=2.15.0,<2.18.0  # Compatible with Python 3.10
numpy>=1.23.5  # Allow newer versions
Pillow>=10.0.0
requests==2.28.1
python-dotenv==0.21.0
gunicorn==20.1.0
setuptools>=68.0.0
pytest==7.4.3
werkzeug>=2.2.3
